[#176191528] refactoring
