+++ ts/features/bonus/bpd/screens/details/components/summary/textualSummary/ClosedTextualSummary.tsx
@@ -1,0 +1,98 @@
+import * as React from "react";
+import { InfoBox } from "../../../../../../../../components/box/InfoBox";
+import { Body } from "../../../../../../../../components/core/typography/Body";
+import I18n from "../../../../../../../../i18n";
+import { dateToAccessibilityReadableFormat } from "../../../../../../../../utils/accessibility";
+import { BpdPeriod } from "../../../../../store/actions/periods";
+import { TextualSummary } from "./TextualSummary";
+
+type Props = React.ComponentProps<typeof TextualSummary>;
+
+/**
+ * We await receipt of the latest transactions to consolidate the count
+ * @param props
+ * @constructor
+ */
+const GracePeriod = (props: { period: BpdPeriod }) => (
+  <InfoBox iconName={"io-hourglass"}>
+    <Body testID={"gracePeriod"}>
+      {I18n.t(
+        "bonus.bpd.details.components.transactionsCountOverview.gracePeriodBody",
+        {
+          date: dateToAccessibilityReadableFormat(props.period.endDate)
+        }
+      )}
+    </Body>
+  </InfoBox>
+);
+
+/**
+ * The user doesn't receive the amount (not enough transactions for the closed period)
+ * @param props
+ * @constructor
+ */
+const KO = (props: Props) => (
+  <InfoBox iconName={"io-sad"}>
+    <Body testID={"closedPeriodKO"}>
+      {I18n.t(
+        "bonus.bpd.details.components.transactionsCountOverview.closedPeriodKOBody",
+        {
+          transactions: props.period.minTransactionNumber,
+          amount: props.period.amount.totalCashback
+        }
+      )}
+    </Body>
+  </InfoBox>
+);
+
+/**
+ * The user will receive the refund!
+ * @param props
+ * @constructor
+ */
+const OK = (props: Props) => (
+  <InfoBox iconName={"io-happy"}>
+    <Body testID={"closedPeriodOK"}>
+      {I18n.t(
+        "bonus.bpd.details.components.transactionsCountOverview.closedPeriodOKBody",
+        {
+          name: props.name,
+          amount: props.period.amount.totalCashback
+        }
+      )}
+      {/* If the max amount is reached, inform the user */}
+      {props.period.amount.totalCashback >= props.period.maxPeriodCashback
+        ? ": " +
+          I18n.t(
+            "bonus.bpd.details.components.transactionsCountOverview.closedPeriodMaxAmount"
+          )
+        : "!"}
+    </Body>
+  </InfoBox>
+);
+
+/**
+ * Return a textual summary for a closed period
+ * @param props
+ * @constructor
+ */
+export const ClosedTextualSummary: React.FunctionComponent<Props> = props => {
+  const today = new Date();
+  const endGracePeriod = new Date(props.period.endDate);
+  endGracePeriod.setDate(
+    props.period.endDate.getDate() + props.period.gracePeriod
+  );
+  // we are still in the grace period and warns the user that some transactions
+  // may still be pending
+  if (today <= endGracePeriod && today >= props.period.endDate) {
+    return <GracePeriod {...props} />;
+  }
+  // not enough transaction to receive the cashback
+  if (
+    props.period.amount.transactionNumber < props.period.minTransactionNumber
+  ) {
+    return <KO {...props} />;
+  }
+  // Congratulation! cashback received
+  return <OK {...props} />;
+};
